pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "remote_job_manager"
version = "0.0.1"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "A command-line tool for managing remote jobs and containerized experiments."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]",
    "pyyaml",
    "rich",
]

[project.urls]
"Homepage" = "https://github.com/example/remote-job-manager"
"Bug Tracker" = "https://github.com/example/remote-job-manager/issues"

[project.scripts]
job-manager = "remote_job_manager.cli:app"
